requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1